refactor ExtHostDocumentModel to ready for node_module consumption
